requirements.txt
absl-py>=1.4.0
accelerate>=1.10.0,<2.0
aiohttp>=3.8.0,<4.0
aiosignal>=1.3.0,<2.0
argon2-cffi>=21.3.0,<22.0
argon2-cffi-bindings>=21.2.0,<22.0
asttokens>=2.2.0,<3.0
astunparse>=1.6.0,<2.0
async-timeout>=4.0.0,<5.0
attrs>=23.1.0,<24.0
audioread>=3.0.1,<4.0
av>=15.0.0,<16.0
babel>=2.17.0,<3.0
backcall>=0.2.0,<1.0
beautifulsoup4>=4.12.0,<5.0
bleach>=6.0.0,<7.0
blinker>=1.9.0,<2.0
blis>=0.7.10,<1.0
Brotli>=1.1.0,<2.0
cachetools>=5.3.0,<6.0
catalogue>=2.0.9,<3.0
certifi>=2023.7.22
cffi>=1.15.0,<2.0
charset-normalizer>=3.2.0,<4.0
click>=8.1.0,<9.0
cloudpickle>=2.2.0,<3.0
cmake>=3.27.0
colorama>=0.4.6,<1.0
comm>=0.1.4,<1.0
confection>=0.1.0,<1.0
contourpy>=1.1.0,<2.0
csvw>=3.5.0,<4.0
cycler>=0.11.0,<1.0
cymem>=2.0.7,<3.0
Cython>=3.0.0,<4.0
dask>=2023.3.0,<2024.0
debugpy>=1.6.0,<2.0
decorator>=5.1.0,<6.0
defusedxml>=0.7.0,<1.0
distributed>=2023.3.0,<2024.0
dm-tree>=0.1.8,<1.0
einops>=0.6.0,<1.0
exceptiongroup>=1.1.0,<2.0
execnet>=2.0.0,<3.0
executing>=1.2.0,<2.0
expecttest>=0.1.3,<1.0
fastjsonschema>=2.18.0,<3.0
filelock>=3.12.0,<4.0
Flask>=3.0.0,<4.0
flask-cors>=6.0.0,<7.0
fonttools>=4.42.0,<5.0
frozenlist>=1.4.0,<2.0
fsspec>=2023.6.0,<2024.0
gast>=0.5.0,<1.0
google-auth>=2.22.0,<3.0
google-auth-oauthlib>=0.4.6,<1.0
grpcio>=1.56.0,<2.0
hf-xet>=1.1.8,<2.0
huggingface-hub>=0.34.0,<1.0
hypothesis>=6.138.0,<7.0
idna>=3.4,<4.0
imageio>=2.37.0,<3.0
imageio-ffmpeg>=0.6.0,<1.0
importlib-metadata>=6.8.0,<7.0
iniconfig>=2.0.0,<3.0
ipykernel>=6.25.0,<7.0
ipython>=8.14.0,<9.0
ipython-genutils>=0.2.0,<1.0
isodate>=0.7.0,<1.0
itsdangerous>=2.2.0,<3.0
jedi>=0.19.0,<1.0
Jinja2>=3.1.0,<4.0
joblib>=1.3.0,<2.0
json5>=0.9.0,<1.0
jsonschema>=4.18.0,<5.0
jsonschema-specifications>=2023.7.0,<2024.0
jupyter_client>=8.3.0,<9.0
jupyter_core>=5.3.0,<6.0
jupyterlab>=2.3.0,<4.0
jupyterlab-pygments>=0.2.0,<1.0
jupyterlab-server>=1.2.0,<3.0
jupytext>=1.15.0,<2.0
kenlm>=0.3.0,<1.0
kiwisolver>=1.4.0,<2.0
langcodes>=3.3.0,<4.0
language-tags>=1.2.0,<2.0
librosa>=0.9.2,<1.0
llvmlite>=0.40.0,<1.0
locket>=1.0.0,<2.0
Markdown>=3.4.0,<4.0
markdown-it-py>=3.0.0,<4.0
MarkupSafe>=2.1.0,<3.0
matplotlib>=3.7.0,<4.0
matplotlib-inline>=0.1.6,<1.0
mdit-py-plugins>=0.4.0,<1.0
mdurl>=0.1.2,<1.0
mistune>=3.0.0,<4.0
mock>=5.1.0,<6.0
mpmath>=1.3.0,<2.0
msgpack>=1.0.0,<2.0
multidict>=6.0.0,<7.0
murmurhash>=1.0.0,<2.0
mutagen>=1.47.0,<2.0
nbclient>=0.8.0,<1.0
nbconvert>=7.7.0,<8.0
nbformat>=5.9.0,<6.0
nest-asyncio>=1.5.0,<2.0
networkx>=2.6.0,<4.0
ninja>=1.11.0,<2.0
notebook>=6.4.0,<7.0
numba>=0.57.0,<1.0
numpy>=1.22.0,<2.0
oauthlib>=3.2.0,<4.0
onnx>=1.14.0,<2.0
opencv-python>=4.7.0,<5.0
packaging>=23.1,<24.0
pandas>=1.5.0,<3.0
pandocfilters>=1.5.0,<2.0
parso>=0.8.0,<1.0
partd>=1.4.0,<2.0
pathy>=0.10.0,<1.0
pexpect>=4.8.0,<5.0
phonemizer>=3.3.0,<4.0
pickleshare>=0.7.0,<1.0
pillow>=11.0.0,<12.0
platformdirs>=3.10.0,<4.0
pluggy>=1.2.0,<2.0
ply>=3.11,<4.0
pooch>=1.7.0,<2.0
preshed>=3.0.0,<4.0
prettytable>=3.8.0,<4.0
prometheus-client>=0.17.0,<1.0
prompt-toolkit>=3.0.0,<4.0
protobuf>=4.21.0,<5.0
psutil>=5.9.0,<6.0
ptyprocess>=0.7.0,<1.0
pure-eval>=0.2.0,<1.0
pyarrow>=14.0.0,<15.0
pyasn1>=0.5.0,<1.0
pyasn1-modules>=0.3.0,<1.0
pybind11>=2.11.0,<3.0
pycocotools>=2.0.7,<3.0
pycparser>=2.21,<3.0
pycryptodomex>=3.23.0,<4.0
pyctcdecode>=0.5.0,<1.0
pydantic>=1.10.0,<2.0
Pygments>=2.16.0,<3.0
pygtrie>=2.5.0,<3.0
pyparsing>=3.0.0,<4.0
pytest>=7.4.0,<8.0
pytest-flakefinder>=1.1.0,<2.0
pytest-rerunfailures>=12.0,<13.0
pytest-shard>=0.1.2,<1.0
pytest-xdist>=3.3.0,<4.0
python-dateutil>=2.8.0,<3.0
python-hostlist>=1.23.0,<2.0
pytz>=2023.3,<2025.0
PyYAML>=6.0.0,<7.0
pyzmq>=25.1.0,<26.0
rdflib>=7.1.0,<8.0
referencing>=0.30.0,<1.0
regex>=2023.6.0,<2025.0
requests>=2.32.0,<3.0
requests-oauthlib>=1.3.0,<2.0
resampy>=0.4.0,<1.0
rfc3986>=1.5.0,<2.0
rpds-py>=0.9.0,<1.0
rsa>=4.9,<5.0
safetensors>=0.6.0,<1.0
scikit-learn>=1.2.0,<2.0
scipy>=1.11.0,<2.0
segments>=2.3.0,<3.0
Send2Trash>=1.8.0,<2.0
six>=1.16.0,<2.0
smart-open>=6.3.0,<7.0
sortedcontainers>=2.4.0,<3.0
soundfile>=0.12.0,<1.0
soupsieve>=2.4.0,<3.0
soxr>=0.5.0,<1.0
spacy>=3.6.0,<4.0
spacy-legacy>=3.0.12,<4.0
spacy-loggers>=1.0.4,<2.0
sphinx-glpi-theme>=0.3,<1.0
srsly>=2.4.0,<3.0
stack-data>=0.6.0,<1.0
supervision>=0.26.0,<1.0
sympy>=1.14.0,<2.0
tabulate>=0.9.0,<1.0
tblib>=2.0.0,<3.0
tensorboard>=2.9.0,<3.0
tensorboard-data-server>=0.6.1,<1.0
tensorboard-plugin-wit>=1.8.0,<2.0
terminado>=0.17.0,<1.0
thinc>=8.1.0,<9.0
threadpoolctl>=3.2.0,<4.0
thriftpy2>=0.4.0,<1.0
tinycss2>=1.2.0,<2.0
tokenizers>=0.21.0,<1.0
toml>=0.10.0,<1.0
tomli>=2.0.0,<3.0
toolz>=0.12.0,<1.0
torch>=2.8.0,<3.0
torchaudio>=2.8.0,<3.0
torchdata>=0.7.0a0,<1.0
torchtext>=0.16.0a0,<1.0
torchvision>=0.16.0a0,<1.0
tornado>=6.3.0,<7.0
tqdm>=4.65.0,<5.0
traitlets>=5.9.0,<6.0
transformers>=4.55.0,<5.0
treelite>=3.2.0,<4.0
treelite-runtime>=3.2.0,<4.0
triton>=3.4.0,<4.0; sys_platform == "linux"
typer>=0.9.0,<1.0
types-dataclasses>=0.6.0,<1.0
typing_extensions>=4.15.0,<5.0
uritemplate>=4.2.0,<5.0
urllib3>=2.5.0,<3.0
wasabi>=1.1.0,<2.0
wcwidth>=0.2.0,<1.0
webencodings>=0.5.0,<1.0
websockets>=15.0.0,<16.0
Werkzeug>=3.1.0,<4.0
xdoctest>=1.0.0,<2.0
xgboost>=1.7.0,<2.0
yarl>=1.9.0,<2.0
yt-dlp>=2025.8.0,<2026.0
zict>=3.0.0,<4.0
zipp>=3.16.0,<4.0
deep-sort-realtime